requirements.txt
Django==4.2.4
Pillow==10.0.0
python-decouple==3.8
bcrypt==4.0.1